+++ modular-core/src/lib.rs
@@ -1,1 +1,5 @@
-pub mod core;
+pub mod error;
+pub mod module;
+pub mod modular;
+pub mod request;
+pub mod response;
