+++ modular-sys/src/core/mod.rs
@@ -4,8 +4,8 @@ use crate::dll::ModuleError;
 use bytes::Bytes;
 use futures_util::future::BoxFuture;
 use futures_util::stream::BoxStream;
-use modular_core::core::error::ModuleError;
-use modular_core::core::module::Module;
+use modular_core::error::ModuleError;
+use modular_core::module::Module;
 use std::future::Future;
 
 pub type BoxModule = Box<dyn Module<Future = BoxFuture<'static, Result<Bytes, ModuleError>>>>;
